requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dnspython==2.6.1
email_validator==2.2.0
emit==0.4.0
fastapi==0.114.1
fastapi-cli==0.0.5
Flask==3.0.3
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
fonttools==4.54.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
numpy==2.1.2
orjson==3.10.7
packaging==24.1
pillow==11.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.9.1
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23.3
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.9
python-socketio==5.11.4
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
setuptools==74.1.2
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.38.5
sympy==1.13.3
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.0.1
Werkzeug==3.0.4
wsproto==1.2.0
